refactor LSP server setup
Move back to a list iteration approach, which works better for the vast
majority of servers. Those that need individual configuration will still
get the default capabilities and on_attach callback, unless the override
it. This is a nicer way than it was happening before too.

put all LSP config in the same Lua block